small bugs corrected

+++ server/app/gateways/classic-mode/classic-mode.gateway.ts
@@ -86,6 +86,7 @@ export class ClassicModeGateway implements OnGatewayConnection, OnGatewayDisconn
     @SubscribeMessage(ClassicModeEvents.AbortGameCreation)
     abortGameCreation(socket: Socket): void {
         const gameRoom = this.classicModeService.getRoom(socket.id);
+        if (!gameRoom) return;
         this.logger.log(`Game creation aborted: ${gameRoom.userGame.gameData.gameForm.name}`);
         this.classicModeService.deleteRoom(socket.id);
         this.server.emit(ClassicModeEvents.GameDeleted, gameRoom.userGame.gameData.gameForm.name);
@@ -95,12 +96,11 @@ export class ClassicModeGateway implements OnGatewayConnection, OnGatewayDisconn
     @SubscribeMessage(ClassicModeEvents.LeaveGame)
     leaveGame(socket: Socket, playerInfo: { roomId: string; username: string }): void {
         const gameRoom = this.classicModeService.getRoom(playerInfo.roomId);
-        if (gameRoom) {
-            this.logger.log(`${playerInfo.username} left the game: ${gameRoom.userGame.gameData.gameForm.name}`);
-            gameRoom.userGame.potentialPlayers = gameRoom.userGame.potentialPlayers.filter((player) => player !== playerInfo.username);
-            this.classicModeService.setRoom(gameRoom);
-            this.server.to(gameRoom.roomId).emit(ClassicModeEvents.GameInfo, gameRoom);
-        }
+        if (!gameRoom) return;
+        this.logger.log(`${playerInfo.username} left the game: ${gameRoom.userGame.gameData.gameForm.name}`);
+        gameRoom.userGame.potentielPlayers = gameRoom.userGame.potentielPlayers.filter((player) => player !== playerInfo.username);
+        this.classicModeService.setRoom(gameRoom);
+        this.server.to(gameRoom.roomId).emit(ClassicModeEvents.GameInfo, gameRoom);
     }
 
     @SubscribeMessage(ClassicModeEvents.RejectPlayer)
@@ -108,7 +108,7 @@ export class ClassicModeGateway implements OnGatewayConnection, OnGatewayDisconn
         const gameRoom = this.classicModeService.getRoom(playerInfo.roomId);
         if (gameRoom) {
             this.logger.log(`${playerInfo.roomId} rejected from game: ${gameRoom.userGame.gameData.gameForm.name}`);
-            gameRoom.userGame.potentialPlayers = gameRoom.userGame.potentialPlayers.filter((player) => player !== playerInfo.username);
+            gameRoom.userGame.potentielPlayers = gameRoom.userGame.potentielPlayers.filter((player) => player !== playerInfo.username);
             this.classicModeService.setRoom(gameRoom);
             this.server.to(gameRoom.roomId).emit(ClassicModeEvents.PlayerRejected, gameRoom);
         }
@@ -119,7 +119,7 @@ export class ClassicModeGateway implements OnGatewayConnection, OnGatewayDisconn
         const gameRoom = this.classicModeService.getRoom(playerInfo.roomId);
         if (gameRoom) {
             this.logger.log(`${playerInfo.username} accepted in game:  ${gameRoom.userGame.gameData.gameForm.name}`);
-            gameRoom.userGame.potentialPlayers = [];
+            gameRoom.userGame.potentielPlayers = [];
             gameRoom.userGame.username2 = playerInfo.username;
             gameRoom.started = true;
             this.classicModeService.setRoom(gameRoom);
@@ -157,6 +157,7 @@ export class ClassicModeGateway implements OnGatewayConnection, OnGatewayDisconn
     }
 
     cancelDeletedGame(gameName: string): void {
+        this.logger.log(`Game canceled: ${gameName}`);
         this.server.emit(ClassicModeEvents.GameCanceled, gameName);
     }
 }
